Raise (new) DataCouldNotBeParsed exception when JSON invalid
Python's inbuilt json module is used to parse JSON in requests and
responses before it's checked against the schema. The json module
only raises ValueError in cases where it is unable to parse the
data, e.g.

    import json
    json.loads('{"not valid": [1,]}'
    >>> ValueError: No JSON object could be decoded

If this happens the ValueError has been allowed to bubble - but
as ValueError can occur in many situations it's not specific enough
to programmatically tell that something has done wrong because the
JSON is invalid (unless you examine the message or stack trace).

This catches ValueError around json.loads in the Request/Response
data property and instead raises a new exception: DataCouldNotBeParsed.
As this is a specific exception for this case it makes it easy to
catch and handle seperately.

NOTE: Request can also parse application/x-www-form-urlencoded data,
      but urlparse.parse_qsl doesn't currently raise exceptions as
      strict_parsing is false by default (so it simply returns empty
      arrays instead of raising exceptions against bad data).